bluefox2: move header file to include
