+++ drivers/bluefox2/include/bluefox2/camera.h
